Fix WP7 bug
The fact that nobody has caught this until now tells me how seriously I should take WP7 development

+++ ReactiveUI/RxApp.cs
@@ -558,7 +558,7 @@ namespace ReactiveUI
     public sealed class RefcountDisposeWrapper
     {
         IDisposable _inner;
-        long refCount = 1;
+        int refCount = 1;
 
         public RefcountDisposeWrapper(IDisposable inner) { _inner = inner; }
